security api & symptoms solutions & more tweaks

+++ app/api/delete/route.ts
@@ -1,4 +1,4 @@
-import { catchAdminAuth } from '@/lib/utils';
+import { isAdminAuth, sendAccessDenied } from '@/lib/utils';
 import { sql } from '@vercel/postgres';
 import { NextRequest, NextResponse } from 'next/server';
 
@@ -9,19 +9,21 @@ interface PostRequestBody {
 
 export async function POST(request: NextRequest) {
     try {
-        catchAdminAuth(request);
-        const requestBody: PostRequestBody = await request.json();
+        if (!await isAdminAuth(request)) return sendAccessDenied();
+        else {
 
-        if (!requestBody.id) {
-            return NextResponse.json({ message: "Invalid body, missing required properties" }, { status: 400 });
+            const requestBody: PostRequestBody = await request.json();
+
+            if (!requestBody.id) {
+                return NextResponse.json({ message: "Invalid body, missing required properties" }, { status: 400 });
+            }
+            const { id } = requestBody;
+
+            const myQuery = sql`DELETE FROM thoracic_posts WHERE id =${id}`;
+
+            const res = await myQuery;
+            return NextResponse.json({ message: "Post removed successfully", data: res });
         }
-        const { id } = requestBody;
-
-        const myQuery = sql`DELETE FROM thoracic_posts WHERE id =${id}`;
-
-        const res = await myQuery;
-
-        return NextResponse.json({ message: "Post removed successfully", data: res });
 
     } catch (error) {
         console.error(error);
